Simplify and clean example
