+++ fetchmap/legend.css
@@ -12,11 +12,11 @@
 
 .legend-container {
   margin: 1rem;
-  width: 180px;
-  max-height: calc(100% - 3rem);
-  overflow-y: auto;
+  width: 220px;
+  max-height: min(80vh, 500px);
+  overflow-y: scroll;
   background: #fdfdfe;
-  padding: 0.75rem;
+  padding: 1.25rem;
   border-radius: 4px;
   box-shadow: 2px 2px 6px #30303099;
   z-index: 1000;
@@ -25,9 +25,29 @@
   pointer-events: auto;
 }
 
+/* Custom Scrollbar for Webkit browsers (Chrome, Safari) - Overriding global hide */
+.legend-container::-webkit-scrollbar {
+  width: 8px !important;
+  display: block !important;
+}
+
+.legend-container::-webkit-scrollbar-track {
+  background: #f1f1f1 !important;
+  border-radius: 10px;
+}
+
+.legend-container::-webkit-scrollbar-thumb {
+  background: #c1c1c1 !important;
+  border-radius: 10px;
+}
+
+.legend-container::-webkit-scrollbar-thumb:hover {
+  background: #a1a1a1 !important;
+}
+
 .legend-layer {
-  margin-bottom: 8px;
-  padding-bottom: 8px;
+  margin-bottom: 12px; /* Increased margin-bottom */
+  padding-bottom: 12px; /* Increased padding-bottom */
   border-bottom: 1px solid #eee;
   position: relative;
 }
@@ -42,7 +62,7 @@
 .legend-title-container {
   display: flex;
   align-items: center;
-  margin-bottom: 6px;
+  margin-bottom: 8px; /* Increased from 6px */
 }
 
 .legend-title {
@@ -79,14 +99,14 @@
 .legend-range {
   display: flex;
   align-items: center;
-  margin-bottom: 4px;
+  margin-bottom: 6px; /* Increased from 4px */
 }
 
 .legend-color-swatch {
   width: 12px;
   height: 12px;
   border-radius: 50%;
-  margin-right: 6px;
+  margin-right: 8px; /* Increased from 6px */
   flex-shrink: 0;
 }
 
@@ -109,68 +129,8 @@
 /* Class for the new details container, if specific styling is needed */
 .legend-layer-details {
   padding-left: 2px;
-  margin-top: 4px;
-  padding-left: 20px;
-}
-
-/* Drag Handler Styles */
-.legend-drag-handler {
-  display: inline-block;
-  width: 12px;
-  height: 20px;
-  margin-right: 6px;
-  cursor: grab;
-  font-family: 'sans-serif';
-  font-size: 16px;
-  line-height: 0.8;
-  text-align: center;
-  color: #aaa;
-  user-select: none;
-  letter-spacing: -2px;
-  overflow: hidden;
-  white-space: nowrap;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-  align-items: center;
-}
-
-.legend-drag-handler::before {
-  content: '\2807\2807';
-  display: none;
-}
-
-.legend-drag-handler span {
-  display: block;
-  line-height: 0.5;
-  font-size: 10px;
-}
-
-/* Drag and Drop Styles */
-.legend-layer-dragging {
-  opacity: 0.5;
-  border: 1px dashed #007bff;
-  background-color: #f0f8ff;
-}
-
-.legend-layer-dragging .legend-drag-handler {
-  cursor: grabbing;
-}
-
-/* .legend-layer-dragover { */ /* Old style, remove or repurpose if needed */
-/* border-top: 2px solid #007bff; */
-/* } */
-
-.legend-drop-above {
-  border-top: 2px solid #007bff !important; /* Blue border on top */
-  box-shadow: 0 -2px 4px -2px #007bff40 inset; /* Optional inner shadow */
-  background-color: rgba(0, 123, 255, 0.1) !important; /* Light blue background highlight */
-}
-
-.legend-drop-below {
-  border-bottom: 2px solid #007bff !important; /* Blue border on bottom */
-  box-shadow: 0 2px 4px -2px #007bff40 inset; /* Optional inner shadow */
-  background-color: rgba(0, 123, 255, 0.1) !important; /* Light blue background highlight */
+  margin-top: 6px; /* Increased from 4px */
+  padding-left: 24px; /* Increased from 20px */
 }
 
 /* Switch Toggle Styles */
